% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate simulation data}
\usage{
gendata(
  n,
  size = 2,
  rho = 0,
  binary_cnf = NULL,
  cont_cnf = NULL,
  mu = NULL,
  sigma = NULL,
  randomized = FALSE
)
}
\arguments{
\item{n}{sample size.}

\item{size}{number of categories in Y, default is 2.}

\item{rho}{logistic distribution informative prior, default is 0.}

\item{binary_cnf}{number of binary confounders, NULL means none.}

\item{cont_cnf}{number of continuous confounders, NUL means none.}

\item{mu}{}

\item{sigma}{}

\item{randomized}{simulate an randomized controlled trial (i.e., probability of treatment is 0.5)?}
}
\value{
A \code{data.table} of the simulated data.
}
\description{
Generate simulation data
}
\author{
Nicholas Williams and Ivan Dìaz
}
