% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{partition}
\alias{partition}
\title{Run simulations in parallel on a computing cluster.}
\usage{
partition(task)
}
\arguments{
\item{task}{A list containing meta information for a simulation task.}
}
\value{
Saves results files.
}
\description{
Run simulations in parallel on a computing cluster.
}
\author{
Nicholas Williams and Iván Díaz
}
