% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Simulate the bias of TMLE and parametric G-computation.}
\usage{
bias(
  context = c("binary", "ordinal", "tte"),
  seed,
  n,
  reps,
  size,
  binary_cnf,
  cont_cnf,
  randomized = FALSE,
  parametric = FALSE
)
}
\arguments{
\item{context}{Outcome type, valid options are "binary", "ordinal", and "tte."}

\item{seed}{Seed for reproducibility.}

\item{n}{The sample size.}

\item{reps}{How many datasets should be generated from the DGM.}

\item{size}{Number of categories in Y, ignored if context is "binary".}

\item{binary_cnf}{Number of binary confounders.}

\item{cont_cnf}{Size of a continuous confounder.}

\item{randomized}{Simulate an randomized controlled trial (i.e., probability of treatment is 0.5)?}

\item{parametric}{Estimate TMLE using just GLM? Only used with \code{context = "binary"}.}
}
\value{
A list of the true value, the result from parametric G-computation,
and the result from TMLE.
}
\description{
Simulate the bias of TMLE and parametric G-computation.
}
\author{
Nicholas Williams and Iván Díaz
}
