% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate simulation data}
\usage{
gendata(
  n,
  size = 2,
  rho = 0,
  binary_cnf,
  cont_cnf,
  mu = NULL,
  sigma = NULL,
  randomized = FALSE
)
}
\arguments{
\item{n}{The sample size.}

\item{size}{Number of categories in Y, default is 2.}

\item{rho}{Logistic distribution informative prior, default is 0.}

\item{binary_cnf}{Number of binary confounders.}

\item{cont_cnf}{Size of a continuous confounder.}

\item{mu}{}

\item{sigma}{}

\item{randomized}{Simulate an randomized controlled trial (i.e., probability of treatment is 0.5)?}
}
\value{
A \code{data.table} of the simulated data.
}
\description{
Generate simulation data
}
\author{
Nicholas Williams and Ivan Dìaz
}
