% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate simulation data}
\usage{
gendata(
  n,
  reps,
  size = 2,
  rho = 0,
  binary_cnf,
  cont_cnf,
  randomized = FALSE,
  seed
)
}
\arguments{
\item{n}{The sample size.}

\item{reps}{How many datasets should be generated from the DGM.}

\item{size}{Number of categories in Y, default is 2.}

\item{rho}{Logistic distribution informative prior, default is 0.}

\item{binary_cnf}{Number of binary confounders.}

\item{cont_cnf}{Size of a continuous confounder.}

\item{randomized}{Simulate an randomized controlled trial (i.e., probability of treatment is 0.5)?}

\item{seed}{Seed for reproducibility.}
}
\value{
A \code{data.table} of the simulated data.
}
\description{
Generate simulation data
}
\author{
Nicholas Williams and Iván Díaz
}
